Refactored and moved files around to make this more maintainable. Output is binary equivalent to before the change.
