+++ tools/Unicode/CaseInsensitive/.gitignore
@@ -1,0 +1,5 @@
+node_modules/
+*.js
+*.map
+*.txt
+out/
