+++ tools/Unicode/CaseInsensitive/tsconfig.json
@@ -2,6 +2,7 @@
     "compilerOptions": {
         "target": "es6",
         "module": "amd",
-        "sourceMap": true
+        "sourceMap": true,
+        "outFile": "mappings.js"
     }
 }
