+++ tools/Unicode/TableToEquiv/table-to-equiv.js
